requirements.txt
contextlib2==0.5.4
Django==1.10.1
django-webpack-loader==0.3.3
django-filter==1.0.2
djangorestframework==3.4.6
funcsigs==1.0.2
future==0.15.0
mock==2.0.0
pbr==1.10.0
psycopg2==2.6.2
py==1.4.31
pytest==3.0.6
pytest-django==3.1.2
PyYAML==3.12
requests==2.7.0
requests-mock==0.6.0
six==1.9.0
sodapy==1.4.0
vcrpy==1.10.5
wrapt==1.10.8
